# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Liberty Mutual Dev API",
		"description": "",
		"contact": {
			"name": "Liberty Mutual",
			"url": "https://lmelg-api.mobileprogramming.net"
		}
	},
	"host": "lmelg-api.mobileprogramming.net",
	"basePath": "/",
	"schemes": [
		"https",
    "http"
	],
	"securityDefinitions": {
		"JWT": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header",
			"description": "Enter valid API authorization token"
		}
	},
...
	"paths": {
...
    "/user/resetPassword": {
      "post": {
        "tags": [
          "User API"
        ],
        "summary": "Reset Password.",
        "description": "To Reset Password",
        "operationId": "resetPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "in": "body",
            "name": "body",           
            "required": true,
            "schema": {
              "$ref": "#/definitions/resetPassword"
            }
          }      
      ] 
      ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/user/resetForgotPassword": {
      "post": {
        "tags": [
          "User API"
        ],
        "summary": "Reset User.",
        "description": "To reset user password into account",
        "operationId": "resetForgotPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",           
            "required": true,
            "schema": {
              "$ref": "#/definitions/resetForgotPassword"
            }
          }      
      ] 
      ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/user/updateProfile": {
      "put": {
        "tags": [
          "User API"
        ],
        "summary": "Update Profile.",
        "description": "To update Profile",
        "operationId": "updateProfile",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "in": "body",
            "name": "body",           
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          }      
      ] 
      ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/user/getProfile": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get User Profile.",
        "description": "To getProfile list",
        "operationId": "getProfile",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          }               
        ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },   
    "/user/getProduct": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get Additional Voluntary Benefits.",
        "description": "To Additional Voluntary Benefits",
        "operationId": "getProduct",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "in": "query",
            "name": "categoryId",
            "description": "categoryId number",
            "required": true,           
            "type": "integer"
          }      
      ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/user/getVideo": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get Content Videos.",
        "description": "To Videos list",
        "operationId": "getVideo",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          } ,
          {
            "in": "query",
            "name": "categoryId",
            "description": "categoryId number",
            "required": true,  
            "type": "integer"
          }              
        ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/user/getProgram": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get program, email template and benefits.",
        "description": "To email template list",
        "operationId": "getProgram",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          } ,
          {
            "in": "query",
            "name": "categoryId",
            "description": "categoryId number",
            "required": true,  
            "type": "integer"
          }              
        ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/user/getParentCategory": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get Parent Category.",
        "description": "To Parent Category list",
        "operationId": "getParentCategory",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          }               
        ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }, 
    "/user/getUserList": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get USer list.",
        "description": "To usermanagement list isBlocked(0=>active,1=>Inactive)",
        "operationId": "getUserList",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "in": "query",
            "name": "search",
            "description": "search keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          },   
          {
            "in": "query",
            "name": "page",
            "description": "page numebr",
            "required": true,
            "default":1,
            "schema": {
              "type": "integer"
            }
          }      
      ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/user/getClientList": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get Client list.",
        "description": "To client management list isBlocked(0=>active,1=>Inactive)",
        "operationId": "getClientList",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "in": "query",
            "name": "search",
            "description": "search keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          },   
          {
            "in": "query",
            "name": "page",
            "description": "page numebr",
            "required": true,
            "default":1,
            "schema": {
              "type": "integer"
            }
          }      
      ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/admin/getProfile": {
      "get": {
        "tags": [
          "Admin API"
        ],
        "summary": "Get all type Profile.",
        "description": "To getProfile list",
        "operationId": "getProfile",
        "produces": [
          "application/json"
        ] ,
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          }               
        ] ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/admin/resetForgotPassword": {
      "post": {
        "tags": [
          "Admin API"
        ],
        "summary": "Reset admin.",
        "description": "To reset admin password into account",
        "operationId": "resetForgotPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",           
            "required": true,
            "schema": {
              "$ref": "#/definitions/resetForgotPasswordAdmin"
            }
          }      
      ] 
      ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/admin/addUser": {
      "post": {
        "tags": [
          "Admin API"
        ],
        "summary": "Add User.",
        "description": "To Add user into account (role=>Normal)",
        "operationId": "addUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "description": "JWT token",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "in": "body",
            "name": "body",           
            "required": true,
            "schema": {
              "$ref": "#/definitions/addUser"
            }
          }      
      ] 
      ,
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
    }
  },
  "definitions": {
...
    "updateProfile": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"         
        },
        "lastName": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "resetPassword": {
      "type": "object",
      "properties": {       
        "new_pass": {
          "type": "string"
        },
        "confirm_pass": {
          "type": "string"
        }
      }
    },
...
    "resetForgotPassword": {
      "type": "object",
      "properties": {
        "resetPasswordToken": {
          "type": "string"         
        },
        "new_pass": {
          "type": "string"
        },
        "confirm_pass": {
          "type": "string"
        }
      }
    },
...
    "resetForgotPasswordAdmin": {
      "type": "object",
      "properties": {
        "resetPasswordToken": {
          "type": "string"         
        },
        "new_pass": {
          "type": "string"
        },
        "confirm_pass": {
          "type": "string"
        }
      }
    },
...
    "addUser": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"         
        },
        "lastName": {
          "type": "string"         
        },
        "email": {
          "type": "string"         
        },
        "role": {
          "type": "string",
          "example": "Normal"
        },
        "brokered": {
          "type": "string",
          "example": "Yes"
        },
        "payroll": {
          "type": "string",
          "example": "Yes"
        },
        "url": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/addURL"
          }
        }
      }
    },
    "addURL": {
      "type": "object",
      "properties": {       
        "clientName": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }       
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "message": {
          "type": "string",
          "description": "Error message.",
          "example": "Network error found. Please try after some time."
        }
      }
    }
  },
...
}